Extract commonalities from Address,Email and Phone class and .........
