+++ src/seedu/addressbook/commands/CommandResult.java
@@ -11,7 +11,7 @@ import java.util.Optional;
 public class CommandResult {
 
     /** The feedback message to be shown to the user. Contains a description of the execution result */
-    public final String feedbackToUser;
+    private final String feedbackToUser;
 
     /** The list of persons that was produced by the command */
     private final List<? extends ReadOnlyPerson> relevantPersons;
@@ -32,5 +32,9 @@ public class CommandResult {
     public Optional<List<? extends ReadOnlyPerson>> getRelevantPersons() {
         return Optional.ofNullable(relevantPersons);
     }
+
+	public String getFeedbackToUser() {
+		return feedbackToUser;
+	}
 
 }
